+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Development version
+--------------------------------------
+- Support for dolfinx v0.10
+- Construct pointwise observation operator
+- Discrete state observation data fidelity
+
 Version 0.1.0-rc, released on May 15, 2024
 --------------------------------------